requirements.txt
# --- Backend Framework ---
# For building the API and serving it
fastapi
uvicorn[standard]

# --- Frontend UI ---
# For creating the user-facing chat interface
streamlit

# --- LLM Integration ---
# For connecting to the NL-to-SQL and Summarization models
openai

# --- Database Connection ---
# SQLAlchemy is a toolkit for SQL, and pyodbc is the driver for MS SQL Server
sqlalchemy
pyodbc
pandas

# --- Utilities ---
# For the Streamlit frontend to call the FastAPI backend
requests
# For loading API keys and DB credentials from a .env file
python-dotenv

#Automation Testing
pytest

#Tablename
#tblItemStatistics

#SQL Validator 
sqlparse
re

numpy